* bug-12084-fix-gcc-4-1-errors-because-of-warnings:
  Use the proper types for level and window variables.
  Removed itkTypeMacro to get rid of unneeded virtual function.
  Use cmath for overloaded std::abs function prototypes
  Fixed warnings with gcc 4.1.2 which are treated as errors.

+++ Core/Code/DataManagement/mitkDataStorage.cpp
@@ -373,7 +373,7 @@ mitk::TimeSlicedGeometry::Pointer mitk::DataStorage::ComputeBoundingGeometry3D( 
   {
     minTimeBounds[0] = minimalTime;
     minTimeBounds[1] = minimalTime + minimalIntervallSize;
-    numberOfTimeSteps = (maximalTime-minimalTime)/minimalIntervallSize;
+    numberOfTimeSteps = static_cast<unsigned int>((maximalTime-minimalTime)/minimalIntervallSize);
   }
 
   TimeSlicedGeometry::Pointer timeSlicedGeometry = NULL;
